+++ app/Http/Controllers/Api/CalendarActionController.php
@@ -88,7 +88,7 @@ class CalendarActionController extends Controller
                 'team_member_id' => $data['team_member_id'] ?? null,
             ]),
             'follow-up' => FollowUp::create([
-                'description'    => $data['description'],
+                'title'          => $data['description'],
                 'follow_up_date' => $data['follow_up_date'] ?? null,
                 'team_member_id' => $data['team_member_id'] ?? null,
             ]),